requirements.txt
Flask>=2.0
python-dotenv>=1.0
mysql-connector-python>=8.0
pymysql>=1.0
pandas>=2.0
openpyxl>=3.1
